+++ src/main/java/KoreanChessApplication.java
@@ -1,3 +1,5 @@
+import object.view.GameView;
+
 public class KoreanChessApplication {
     public static void main(String[] args) {
         GameView gameView = new GameView();
