+++ src/main/java/object/Route.java
@@ -1,0 +1,17 @@
+package object;
+
+import java.util.Collections;
+import java.util.List;
+
+public class Route {
+
+    private final List<Coordinate> coordinates;
+
+    public Route(final List<Coordinate> coordinates) {
+        this.coordinates = coordinates;
+    }
+
+    public List<Coordinate> getPositions() {
+        return Collections.unmodifiableList(coordinates);
+    }
+}
