+++ src/main/java/object/strategy/GungMoveStrategy.java
@@ -1,0 +1,21 @@
+package object.strategy;
+
+import object.piece.Pieces;
+import object.Coordinate;
+import object.Route;
+import object.piece.Team;
+
+public class GungMoveStrategy implements MoveStrategy {
+
+    private static final String NOT_IMPLEMENTED_FEATURE = "아직 구현되지 않은 기능입니다.";
+
+    @Override
+    public Route getLegalRoute(Coordinate startCoordinate, Coordinate endCoordinate) {
+        throw new IllegalStateException(NOT_IMPLEMENTED_FEATURE);
+    }
+
+    @Override
+    public Coordinate move(Coordinate destination, Pieces onRoutePieces, Team moveTeam) {
+        throw new IllegalStateException(NOT_IMPLEMENTED_FEATURE);
+    }
+}
