--- src/main/java/strategy/MaMoveStrategy.java
@@ -1,24 +1,0 @@
-package strategy;
-
-import java.util.List;
-import piece.Position;
-import piece.Route;
-
-public class MaMoveStrategy implements MoveStrategy {
-
-    private final List<Route> canMoveDirections = List.of(
-            new Route(List.of(new Position(1, 0), new Position(1, -1))),
-            new Route(List.of(new Position(1, 0), new Position(1, 1))),
-            new Route(List.of(new Position(-1, 0), new Position(-1, -1))),
-            new Route(List.of(new Position(-1, 0), new Position(-1, 1))),
-            new Route(List.of(new Position(0, 1), new Position(1, 1))),
-            new Route(List.of(new Position(0, 1), new Position(-1, 1))),
-            new Route(List.of(new Position(0, -1), new Position(1, 1))),
-            new Route(List.of(new Position(0, -1), new Position(-1, -1)))
-    );
-
-    @Override
-    public Route getLegalRoute(Position startPosition, Position endPosition) {
-        return getLegalRoute(startPosition, endPosition, canMoveDirections);
-    }
-}
