+++ src/main/resources/initialPieces.txt
